Migrate third party posts from collection to data
Collections are a powerful tool but only Jekyll supports them.  I am
thinking about the possibility to move to hugo because I know Go way
better and I am not really thrilled about the idea to learn Ruby.

Data are a supported feature by Hugo and the third party posts are easy
to migrate.